+++ android/src/main/java/com/vanethos/notification_permissions/NotificationPermissionsPlugin.java
@@ -2,74 +2,71 @@ package com.vanethos.notification_permissions;
 
 import android.app.Activity;
 import android.content.Context;
-import android.content.Intent;
-import android.net.Uri;
-import android.os.Build;
-import android.provider.Settings;
-import android.util.Log;
-
-import androidx.core.app.NotificationManagerCompat;
-import io.flutter.plugin.common.MethodCall;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import io.flutter.embedding.engine.plugins.activity.ActivityAware;
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
+import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
 
-public class NotificationPermissionsPlugin implements MethodChannel.MethodCallHandler {
+public class NotificationPermissionsPlugin implements FlutterPlugin, ActivityAware {
   public static void registerWith(Registrar registrar) {
-    final MethodChannel channel =
-        new MethodChannel(registrar.messenger(), "notification_permissions");
-    channel.setMethodCallHandler(new NotificationPermissionsPlugin(registrar));
-  }
+    final NotificationPermissionsPlugin plugin = new NotificationPermissionsPlugin();
+    plugin.onAttachedToEngine(registrar.context(), registrar.messenger());
 
-  private static final String PERMISSION_GRANTED = "granted";
-  private static final String PERMISSION_DENIED = "denied";
-
-  private final Context context;
-
-  private NotificationPermissionsPlugin(Registrar registrar) {
-    this.context = registrar.activity();
-  }
-
-  @Override
-  public void onMethodCall(MethodCall call, MethodChannel.Result result) {
-    if ("getNotificationPermissionStatus".equalsIgnoreCase(call.method)) {
-      result.success(getNotificationPermissionStatus());
-    } else if ("requestNotificationPermissions".equalsIgnoreCase(call.method)) {
-      if (PERMISSION_DENIED.equalsIgnoreCase(getNotificationPermissionStatus())) {
-        if (context instanceof Activity) {
-          // https://stackoverflow.com/a/45192258
-          final Intent intent = new Intent();
-
-          if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
-          // ACTION_APP_NOTIFICATION_SETTINGS was introduced in API level 26 aka Android O
-            intent.setAction(Settings.ACTION_APP_NOTIFICATION_SETTINGS);
-            intent.putExtra(Settings.EXTRA_APP_PACKAGE, context.getPackageName());
-          } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP){
-            intent.setAction("android.settings.APP_NOTIFICATION_SETTINGS");
-            intent.putExtra("app_package", context.getPackageName());
-            intent.putExtra("app_uid", context.getApplicationInfo().uid);
-          } else {
-            intent.setAction(Settings.ACTION_APPLICATION_DETAILS_SETTINGS);
-            intent.addCategory(Intent.CATEGORY_DEFAULT);
-            intent.setData(Uri.parse("package:" + context.getPackageName()));
-          }
-
-          context.startActivity(intent);
-
-          result.success(PERMISSION_DENIED);
-        } else {
-          result.error(call.method, "context is not instance of Activity", null);
-        }
-      } else {
-        result.success(PERMISSION_GRANTED);
-      }
-    } else {
-      result.notImplemented();
+    if (registrar.activity() != null) {
+      plugin.onActivityChanged(registrar.activity());
     }
   }
 
-  private String getNotificationPermissionStatus() {
-    return (NotificationManagerCompat.from(context).areNotificationsEnabled())
-        ? PERMISSION_GRANTED
-        : PERMISSION_DENIED;
+  @Nullable private MethodChannel channel;
+
+  @Nullable private MethodCallHandlerImpl methodCallHandler;
+
+  @Override
+  public void onAttachedToEngine(@NonNull FlutterPluginBinding binding) {
+    onAttachedToEngine(binding.getApplicationContext(), binding.getBinaryMessenger());
+  }
+
+  @Override
+  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
+    if (channel != null) {
+      channel.setMethodCallHandler(null);
+    }
+    channel = null;
+  }
+
+  private void onAttachedToEngine(Context applicationContext, BinaryMessenger messenger) {
+    channel = new MethodChannel(messenger, "notification_permissions");
+    methodCallHandler = new MethodCallHandlerImpl(applicationContext);
+    channel.setMethodCallHandler(methodCallHandler);
+  }
+
+  @Override
+  public void onAttachedToActivity(@NonNull ActivityPluginBinding binding) {
+    onActivityChanged(binding.getActivity());
+  }
+
+  @Override
+  public void onDetachedFromActivityForConfigChanges() {
+    onActivityChanged(null);
+  }
+
+  @Override
+  public void onReattachedToActivityForConfigChanges(@NonNull ActivityPluginBinding binding) {
+    onActivityChanged(binding.getActivity());
+  }
+
+  @Override
+  public void onDetachedFromActivity() {
+    onActivityChanged(null);
+  }
+
+  private void onActivityChanged(@Nullable Activity activity) {
+    if (methodCallHandler != null) {
+      methodCallHandler.setActivity(activity);
+    }
   }
 }